+++ packages/frontend/src/components/MkSignupDialog.rules.stories.impl.ts
@@ -3,7 +3,7 @@ import { expect } from '@storybook/jest';
 import { userEvent, waitFor, within } from '@storybook/testing-library';
 import { StoryObj } from '@storybook/vue3';
 import { onBeforeUnmount } from 'vue';
-import MkSignupServerRules from './MkSignupServerRules.vue';
+import MkSignupServerRules from './MkSignupDialog,rules.vue';
 import { i18n } from '@/i18n';
 import { instance } from '@/instance';
 export const Empty = {
